Application/backend/requirements.txt
bcrypt==4.0.1
click==8.1.3
direnv==2020.12.3
dnspython==2.3.0
env-file==2020.12.3
flake8==6.0.0
Flask~=2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mccabe==0.7.0
mysql==0.0.3
mysqlclient==2.1.1
optional-django==0.1.0
packaging~=23.1
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT~=2.7.0
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy~=2.0.15
typing_extensions==4.5.0
values==2020.12.3
zipp==3.15.0

blinker~=1.6.2
boto3~=1.26.137